Update create_cmds.py
I'm proposing a fix to create_cmds.py since I found a minor bug in line 165 which doesn't remove the excluded dates during image stacking. Instead line 165 removes dates based on the dates not included from the dates to be included parameter on the alosStack.xml.

The image folder contains the ALOS-2 images with dates 191005, 191019, 191102, 191214, 191228.

I've tested the changes and it worked properly.